Use floating-point refresh rate in sys::DisplayMode

+++ src/bstone/src/bstone_sys_display_mode.h
@@ -15,7 +15,7 @@ struct DisplayMode
 {
 	int width;
 	int height;
-	int refresh_rate;
+	float refresh_rate;
 };
 
 } // namespace bstone::sys
